find new ax reps? recognise shape?
Finding "new" axial representations means identifying idempotents in a
known algebra and building a representation around this. The idempotents
can have different fusion rules to before.

To make these results intelligble, we need to be able to recognise the
shape of a representation. This means developing a (good-enough)
algorithm to identify isomorphic (in particular, 2-generated)
subalgebras.